move error banner and refactor tests
