+++ src/pages/PullRequestPage/PullCoverage/ErrorBanner/ErrorBanner.jsx
@@ -114,8 +114,6 @@ function BannerContent({ errorType }) {
       </>
     )
   }
-
-  return null
 }
 
 BannerContent.propTypes = {
